auth_jwt_app_full/requirements.txt
Flask==2.3.2
PyJWT==2.8.0
bcrypt==4.0.1
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
